pyproject.toml
[project]
name = "arq-website"
version = "0.1.0"
description = "ARQ company website"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "jinja2",
    "python-dotenv",
    "pytest",
    # passlib[argon2]
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

# РЕШЕНИЕ КОНФЛИКТА app/alembic:
[tool.setuptools]
# Вариант А: Просто укажи явно пакет app
packages = ["app"]

# Вариант Б: Или используй поиск с исключением alembic
# [tool.setuptools.packages.find]
# where = ["."]
# include = ["app*"]
# exclude = ["alembic*", "tests*", "scripts*", "*.egg-info"]
